[akka-with-client-configs] -  inserted client configs in akka
