+++ akka-bbb-apps/src/universal/conf/application.conf
@@ -27,6 +27,11 @@ akka {
     }
 }
 
+client {
+    clientConfigurationFilePath = "/usr/share/meteor/bundle/programs/server/assets/app/config/settings.yml"
+    clientConfigurationOverrideFilePath = "/etc/bigbluebutton/bbb-html5.yml"
+}
+
 redis {
     host="127.0.0.1"
     port=6379
